+++ docs/changelog.md
@@ -3,6 +3,9 @@
 ## Upcoming version
 
 - Added `fisherman` skill for Elliott
+- Added support for availability (via EPU condition in companion dispositions file)
+- Added translation registry. Translations must have their own `manifest.json` file
+- Added support for optional assets in ContentLoader
 - Added support for EPU conditions in companion event cutscenes
 
 ## 0.15.0 "Tasty Kimchi"